refactor: split dashboard/schemas.py into schemas/ package
