+++ initrunner/dashboard/schemas/__init__.py
@@ -1,0 +1,281 @@
+"""Pydantic response/request models for the dashboard API."""
+
+from initrunner.dashboard.schemas._common import (
+    AgentSlotModel,
+    AgentSlotOption,
+    DeleteResponse,
+    ItemSummary,
+    ModelOption,
+    ProviderModels,
+    ProviderPreset,
+    ProviderStatus,
+    SkillRef,
+    TimelineCostResponse,
+    TimelineEntryResponse,
+    TimelineResponse,
+    TimelineStatsResponse,
+    ValidationIssueResponse,
+)
+from initrunner.dashboard.schemas.agents import (
+    AgentDetail,
+    AgentSummary,
+    RunRequest,
+    RunResponse,
+)
+from initrunner.dashboard.schemas.audit import (
+    AuditRecordResponse,
+    AuditStatsResponse,
+    TopAgentResponse,
+    TriggerStatResponse,
+)
+from initrunner.dashboard.schemas.builder import (
+    BuilderOptionsResponse,
+    EmbeddingOption,
+    EmbeddingWarning,
+    EnvVarStatus,
+    HubSearchResponse,
+    HubSearchResultResponse,
+    HubSeedRequest,
+    SaveRequest,
+    SaveResponse,
+    SeedRequest,
+    SeedResponse,
+    SetEmbeddingProviderRequest,
+    StarterInfo,
+    StartersResponse,
+    TemplateInfo,
+    TemplateSetup,
+    ValidateRequest,
+)
+from initrunner.dashboard.schemas.cost import (
+    AgentCostResponse,
+    CostSummaryResponse,
+    DailyCostResponse,
+    ModelCostResponse,
+)
+from initrunner.dashboard.schemas.flow import (
+    AgentStepResponse,
+    DelegateEventResponse,
+    FlowAgentDetail,
+    FlowBuilderOptionsResponse,
+    FlowDetail,
+    FlowRunRequest,
+    FlowRunResponse,
+    FlowSaveRequest,
+    FlowSaveResponse,
+    FlowSeedRequest,
+    FlowSeedResponse,
+    FlowStatsResponse,
+    FlowSummary,
+    FlowValidateRequest,
+    FlowValidateResponse,
+    FlowYamlSaveRequest,
+    FlowYamlSaveResponse,
+    HealthCheckDetail,
+    PatternInfo,
+    RestartDetail,
+    SinkDetail,
+    SlotAssignment,
+)
+from initrunner.dashboard.schemas.ingest import (
+    AddUrlRequest,
+    IngestDocumentResponse,
+    IngestFileResultResponse,
+    IngestStatsResponse,
+    IngestSummaryResponse,
+)
+from initrunner.dashboard.schemas.mcp import (
+    McpAgentRefResponse,
+    McpHealthResponse,
+    McpHealthSummaryResponse,
+    McpPlaygroundRequest,
+    McpPlaygroundResponse,
+    McpRegistryEntryResponse,
+    McpServerResponse,
+    McpToolResponse,
+)
+from initrunner.dashboard.schemas.memory import (
+    MemoryResponse,
+    SessionDetailResponse,
+    SessionMessageResponse,
+    SessionSummaryResponse,
+)
+from initrunner.dashboard.schemas.skills import (
+    RequirementStatusResponse,
+    SkillAgentRef,
+    SkillContentResponse,
+    SkillContentSaveRequest,
+    SkillContentSaveResponse,
+    SkillCreateRequest,
+    SkillCreateResponse,
+    SkillDeleteBlockedResponse,
+    SkillDetail,
+    SkillSummary,
+    SkillToolSummary,
+)
+from initrunner.dashboard.schemas.system import (
+    DefaultModelResponse,
+    DoctorCheck,
+    DoctorResponse,
+    HealthResponse,
+    ProviderResponse,
+    ProviderStatusResponse,
+    SaveDefaultModelRequest,
+    SaveKeyRequest,
+    SaveKeyResponse,
+    ToolTypeResponse,
+)
+from initrunner.dashboard.schemas.team import (
+    PersonaDetail,
+    PersonaSeedEntry,
+    PersonaSeedModel,
+    PersonaStepResponse,
+    TeamBuilderOptionsResponse,
+    TeamDetail,
+    TeamRunRequest,
+    TeamRunResponse,
+    TeamSaveRequest,
+    TeamSaveResponse,
+    TeamSeedRequest,
+    TeamSeedResponse,
+    TeamSummary,
+    TeamValidateRequest,
+    TeamValidateResponse,
+    TeamYamlSaveRequest,
+    TeamYamlSaveResponse,
+)
+
+__all__ = [
+    # ingest
+    "AddUrlRequest",
+    # cost
+    "AgentCostResponse",
+    # agents
+    "AgentDetail",
+    # _common
+    "AgentSlotModel",
+    "AgentSlotOption",
+    # flow
+    "AgentStepResponse",
+    "AgentSummary",
+    # audit
+    "AuditRecordResponse",
+    "AuditStatsResponse",
+    # builder
+    "BuilderOptionsResponse",
+    "CostSummaryResponse",
+    "DailyCostResponse",
+    # system
+    "DefaultModelResponse",
+    "DelegateEventResponse",
+    "DeleteResponse",
+    "DoctorCheck",
+    "DoctorResponse",
+    "EmbeddingOption",
+    "EmbeddingWarning",
+    "EnvVarStatus",
+    "FlowAgentDetail",
+    "FlowBuilderOptionsResponse",
+    "FlowDetail",
+    "FlowRunRequest",
+    "FlowRunResponse",
+    "FlowSaveRequest",
+    "FlowSaveResponse",
+    "FlowSeedRequest",
+    "FlowSeedResponse",
+    "FlowStatsResponse",
+    "FlowSummary",
+    "FlowValidateRequest",
+    "FlowValidateResponse",
+    "FlowYamlSaveRequest",
+    "FlowYamlSaveResponse",
+    "HealthCheckDetail",
+    "HealthResponse",
+    "HubSearchResponse",
+    "HubSearchResultResponse",
+    "HubSeedRequest",
+    "IngestDocumentResponse",
+    "IngestFileResultResponse",
+    "IngestStatsResponse",
+    "IngestSummaryResponse",
+    "ItemSummary",
+    # mcp
+    "McpAgentRefResponse",
+    "McpHealthResponse",
+    "McpHealthSummaryResponse",
+    "McpPlaygroundRequest",
+    "McpPlaygroundResponse",
+    "McpRegistryEntryResponse",
+    "McpServerResponse",
+    "McpToolResponse",
+    # memory
+    "MemoryResponse",
+    "ModelCostResponse",
+    "ModelOption",
+    "PatternInfo",
+    # team
+    "PersonaDetail",
+    "PersonaSeedEntry",
+    "PersonaSeedModel",
+    "PersonaStepResponse",
+    "ProviderModels",
+    "ProviderPreset",
+    "ProviderResponse",
+    "ProviderStatus",
+    "ProviderStatusResponse",
+    # skills
+    "RequirementStatusResponse",
+    "RestartDetail",
+    "RunRequest",
+    "RunResponse",
+    "SaveDefaultModelRequest",
+    "SaveKeyRequest",
+    "SaveKeyResponse",
+    "SaveRequest",
+    "SaveResponse",
+    "SeedRequest",
+    "SeedResponse",
+    "SessionDetailResponse",
+    "SessionMessageResponse",
+    "SessionSummaryResponse",
+    "SetEmbeddingProviderRequest",
+    "SinkDetail",
+    "SkillAgentRef",
+    "SkillContentResponse",
+    "SkillContentSaveRequest",
+    "SkillContentSaveResponse",
+    "SkillCreateRequest",
+    "SkillCreateResponse",
+    "SkillDeleteBlockedResponse",
+    "SkillDetail",
+    "SkillRef",
+    "SkillSummary",
+    "SkillToolSummary",
+    "SlotAssignment",
+    "StarterInfo",
+    "StartersResponse",
+    "TeamBuilderOptionsResponse",
+    "TeamDetail",
+    "TeamRunRequest",
+    "TeamRunResponse",
+    "TeamSaveRequest",
+    "TeamSaveResponse",
+    "TeamSeedRequest",
+    "TeamSeedResponse",
+    "TeamSummary",
+    "TeamValidateRequest",
+    "TeamValidateResponse",
+    "TeamYamlSaveRequest",
+    "TeamYamlSaveResponse",
+    "TemplateInfo",
+    "TemplateSetup",
+    "TimelineCostResponse",
+    "TimelineEntryResponse",
+    "TimelineResponse",
+    "TimelineStatsResponse",
+    "ToolTypeResponse",
+    "TopAgentResponse",
+    "TriggerStatResponse",
+    "ValidateRequest",
+    "ValidationIssueResponse",
+]
